Add files via upload
Okay, I updated these pretty extensively but I think they have addressed everything and I also caught one more weird copy-paste thing. I updated the questions and vars sheet too since 77 is an acceptable code for the SIBS variable. It's on our list to update the PDF. In UDSv3 it was in the DED, but not on the PDF itself for some reason.